Move send email logic out of view
